+++ src/main/java/roomescape/payment/dto/PaymentConfirmRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto;
+package roomescape.payment.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
