+++ src/main/java/roomescape/payment/service/PaymentClient.java
@@ -1,6 +1,6 @@
-package roomescape.reservation.service;
+package roomescape.payment.service;
 
-import roomescape.reservation.dto.PaymentConfirmRequest;
+import roomescape.payment.dto.PaymentConfirmRequest;
 
 public interface PaymentClient {
     void requestConfirmPayment(PaymentConfirmRequest paymentConfirmRequest);
