+++ src/main/java/roomescape/reservation/dto/MemberReservationAddRequest.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import java.time.LocalDate;
+import roomescape.payment.dto.PaymentConfirmRequest;
 
 public record MemberReservationAddRequest(
         @NotNull(message = "예약 날짜는 필수 입니다.") LocalDate date,
