+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -38,7 +38,8 @@ import roomescape.global.exception.IllegalRequestException;
 import roomescape.global.exception.InternalServerException;
 import roomescape.member.fixture.MemberFixture;
 import roomescape.member.service.MemberService;
-import roomescape.reservation.config.PaymentConfig;
+import roomescape.payment.config.PaymentConfig;
+import roomescape.payment.service.TossPaymentClient;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.domain.ReservationWithWaiting;
